refactor: Bank 클래스 추가 및 역할 분배
