+++ BankManagerConsoleApp/BankManagerConsoleApp/Bank/BankManagerError.swift
@@ -10,4 +10,3 @@ enum BankManagerError: LocalizedError {
         }
     }
 }
-
