+++ BankManagerConsoleApp/BankManagerConsoleApp/main.swift
@@ -1,4 +1,5 @@
 let banker = Banker()
-let bankManager = BankManager(banker: banker)
+let bank = Bank(banker: banker)
+let bankManager = BankManager(bank: bank)
 
 bankManager.openBank()
